+++ 손영석/lotto_simulator/LottoSimulator.java
@@ -1,0 +1,14 @@
+package lotto_simulator;
+
+public class LottoSimulator {
+	static LottoSimulatorModel model = new LottoSimulatorModel();
+	static LottoSimulatorView view = new LottoSimulatorView();
+	static LottoSimulatorController control = new LottoSimulatorController();
+	
+	public static void main(String[] args) {
+		
+		control.mainMenu();
+		
+	}
+
+}
